+++ lib/hcloud/entities/route.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module HCloud
+  class Route < Entity
+    attribute :destination
+    attribute :gateway
+  end
+end