+++ lib/hcloud/entities/subnet.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module HCloud
+  class Subnet < Entity
+    attribute :type
+    attribute :network_zone
+    attribute :gateway
+    attribute :ip_range
+    attribute :vswitch_id
+  end
+end